+++ krylov/lgmres.h
@@ -54,7 +54,7 @@
 
 typedef struct
 {
-   char * (*CAlloc)        ( int count, int elt_size );
+   char * (*CAlloc)        ( size_t count, size_t elt_size );
    int    (*Free)          ( char *ptr );
    int    (*CommInfo)      ( void  *A, int   *my_id, int   *num_procs );
    void * (*CreateVector)  ( void *vector );
@@ -142,7 +142,7 @@ extern "C" {
 
 hypre_LGMRESFunctions *
 hypre_LGMRESFunctionsCreate(
-   char * (*CAlloc)        ( int count, int elt_size ),
+   char * (*CAlloc)        ( size_t count, size_t elt_size ),
    int    (*Free)          ( char *ptr ),
    int    (*CommInfo)      ( void  *A, int   *my_id, int   *num_procs ),
    void * (*CreateVector)  ( void *vector ),
